Add files via upload
Pan Number Validation using postgreSQL

This project contains the pancard numbers which are stored in CSV file which required cleaning, preprocessing, handling missing values, and validation. Therefore, I have used PostgreSQL to complete this task using all the essential concepts of PostgreSQL. This project data was collected from Youtube and build using different logic and approach. All the files are attached to this repository.

This project is build to showcase the skill of PostgreSQL to the recruiters.